refactor: Abstract Factory pattern & subscribe-publish pattern
Changed to Abstract Factory pattern and made sure the subscribe-publish pattern works
