+++ core/src/main/java/com/github/FishMiner/domain/ecs/entityFactories/IEntityFactory.java
@@ -1,7 +1,7 @@
-package com.github.FishMiner.domain.ecs.entityFactories;
-
-import com.badlogic.ashley.core.Entity;
-
-public interface IEntityFactory {
-    Entity createEntity();
-}
+//package com.github.FishMiner.domain.ecs.entityFactories;
+// TODO: REMOVE FILE
+// import com.badlogic.ashley.core.Entity;
+//
+//public interface IEntityFactory {
+//    Entity createEntity();
+//}
