+++ core/src/main/java/com/github/FishMiner/ui/PlayScreen.java
@@ -12,8 +12,8 @@ import com.badlogic.gdx.utils.ScreenUtils;
 import com.github.FishMiner.Configuration;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.Actor;
-import com.github.FishMiner.domain.ecs.entityFactories.impl.FishFactory;
-import com.github.FishMiner.domain.ecs.entityFactories.impl.HookFactory;
+import com.github.FishMiner.domain.ecs.entityFactories.IGameEntityFactory;
+import com.github.FishMiner.domain.ecs.entityFactories.impl.BasicGameEntityFactory;
 import com.github.FishMiner.domain.ecs.systems.AnimationSystem;
 import com.github.FishMiner.domain.ecs.systems.CollisionSystem;
 import com.github.FishMiner.domain.ecs.systems.HookSystem;
@@ -23,8 +23,10 @@ import com.github.FishMiner.domain.ecs.systems.RotationSystem;
 import com.github.FishMiner.domain.events.EventBus;
 import com.github.FishMiner.domain.events.impl.FireHookEvent;
 import com.github.FishMiner.domain.events.impl.FishHitEvent;
+import com.github.FishMiner.domain.events.impl.HookReelingEvent;
 import com.github.FishMiner.domain.listeners.FireHookListener;
 import com.github.FishMiner.domain.listeners.FishStateListener;
+import com.github.FishMiner.domain.listeners.HookStateListener;
 import com.github.FishMiner.ui.controller.InputController;
 
 /**
@@ -40,31 +42,31 @@ public class PlayScreen extends AbstractScreen {
     public void show() {
         super.show();
 
-        // Create a full-width control window for gameplay UI
-        Window controlWindow = new Window("", skin, "border");
-        controlWindow.setWidth(stage.getWidth()); // Full width of the screen
-        controlWindow.defaults().pad(4f);
-        controlWindow.add("Fish Farmer Game").expandX();
-
-        // Create a Menu button that opens the overlay
-        TextButton menuButton = new TextButton("Menu", skin);
-        menuButton.addListener(new ChangeListener() {
-            @Override
-            public void changed(ChangeEvent event, Actor actor) {
-                // Open the PlayMenuScreen overlay
-                PlayMenuScreen playMenuOverlay = new PlayMenuScreen(stage);
-                stage.addActor(playMenuOverlay.getOverlay());
-            }
-        });
-        controlWindow.add(menuButton).pad(4f);
-        controlWindow.pack();
-        // Position the control window at the top of the screen
-        controlWindow.setPosition(0, stage.getHeight() - controlWindow.getHeight());
-        stage.addActor(controlWindow);
+//        // Create a full-width control window for gameplay UI
+//        Window controlWindow = new Window("", skin, "border");
+//        controlWindow.setWidth(stage.getWidth()); // Full width of the screen
+//        controlWindow.defaults().pad(4f);
+//        controlWindow.add("Fish Farmer Game").expandX();
+//
+//        // Create a Menu button that opens the overlay
+//        TextButton menuButton = new TextButton("Menu", skin);
+//        menuButton.addListener(new ChangeListener() {
+//            @Override
+//            public void changed(ChangeEvent event, Actor actor) {
+//                // Open the PlayMenuScreen overlay
+//                PlayMenuScreen playMenuOverlay = new PlayMenuScreen(stage);
+//                stage.addActor(playMenuOverlay.getOverlay());
+//            }
+//        });
+//        controlWindow.add(menuButton).pad(4f);
+//        controlWindow.pack();
+//        // Position the control window at the top of the screen
+//        controlWindow.setPosition(0, stage.getHeight() - controlWindow.getHeight());
+//        stage.addActor(controlWindow);
 
         // Set up event listeners
         EventBus.getInstance().register(FishHitEvent.class, new FishStateListener());
-
+        EventBus.getInstance().register(HookReelingEvent.class, new HookStateListener());
 
         // Initialize the ECS engine, input controller, and sprite batch
         engine = new Engine();
@@ -74,19 +76,21 @@ public class PlayScreen extends AbstractScreen {
         // Add ECS systems
         RotationSystem rotationSystem = new RotationSystem();
         rotationSystem.setHookPosition(200, Configuration.getInstance().getOceanHeight());
+        engine.addSystem(new CollisionSystem());
         engine.addSystem(rotationSystem);
         engine.addSystem(new AnimationSystem());
         engine.addSystem(new MovementSystem());
         engine.addSystem(new RenderingSystem(batch));
-        engine.addSystem(new CollisionSystem());
         engine.addSystem(new HookSystem());
         // Uncomment if needed: engine.addSystem(new HookInputSystem());
 
         // Create and add entities
-        Entity fish = new FishFactory().createEntity();
-        Entity hook = new HookFactory().createEntity();
+        IGameEntityFactory entityFactory = new BasicGameEntityFactory(); // Abstract factory pattern
+        Entity fish = entityFactory.createFish();
+        Entity hook = entityFactory.createHook();
         engine.addEntity(fish);
         engine.addEntity(hook);
+
 
         // Set up input processing with an InputMultiplexer
         InputMultiplexer multiplexer = new InputMultiplexer(controller, stage);
